+++ libs/shared/flagd-core/src/index.ts
@@ -1,4 +1,5 @@
 export * from './lib/flagd-core';
 export * from './lib/feature-flag';
 export * from './lib/storage';
+export * from './lib/options';
 export * from './e2e';